+++ tests/test_serializers.py
@@ -1,6 +1,7 @@
 import copy
 import dataclasses
 import datetime
+import sys
 import typing
 
 from unittest import TestCase